requirements.txt
paho-mqtt==1.5.1
python-statemachine==0.8.0